test: add credits e2e tests
