+++ package.json
@@ -34,6 +34,7 @@
 		"execa": "^9.5.1",
 		"husky": "^8.0.3",
 		"pm2": "^5.4.2",
+		"relative-day-utc": "^1.3.0",
 		"typescript": "~5.2.2"
 	},
 	"scripts": {
