+++ seeds/e2e/user-data.js
@@ -1,10 +1,3 @@
-import { randomUUID } from 'node:crypto';
-
-/**
- * @param { import("knex").Knex } knex
- * @returns { Promise<void> }
- */
-
 export const seed = async (knex) => {
 	const getUser = async () => {
 		return knex('directus_users')
@@ -47,200 +40,4 @@ export const seed = async (knex) => {
 		knex('gp_credits_deductions').delete(),
 		knex('gp_credits').delete(),
 	]);
-
-	await knex('gp_apps').insert([{
-		id: randomUUID(),
-		user_created: user.id,
-		date_created: '2024-09-01 00:00:00',
-		name: 'Sample OAuth App',
-		grants: JSON.stringify([ 'authorization_code', 'refresh_token' ]),
-		redirect_urls: JSON.stringify([ 'http://localhost:13010' ]),
-	}, {
-		id: randomUUID(),
-		user_created: user.id,
-		date_created: '2024-09-01 00:00:00',
-		name: 'Client Credentials App',
-		secrets: JSON.stringify([ 'xhQKGe2+hSCGgAnYYNF6uBNMFJ1YvcpaRzVs+JSpaWw=' ]), // secret: lyrhib7f2dtuh6fzojvupfhh4olkxofd4kibutw6z5guihvz
-		grants: JSON.stringify([ 'globalping_client_credentials', 'refresh_token' ]),
-		redirect_urls: JSON.stringify([ 'http://localhost:13010' ]),
-	}]);
-
-	const probeId = randomUUID();
-	await knex('gp_adopted_probes').insert([{
-		id: probeId,
-		asn: 3302,
-		city: 'Naples',
-		country: 'IT',
-		countryOfCustomCity: 'IT',
-		date_created: '2024-02-22 11:04:30',
-		date_updated: '2024-02-22 11:05:48',
-		ip: '2a02:a319:80f3:8b80:344b:35ff:fee8:a8c',
-		altIps: JSON.stringify([ '213.136.174.80' ]),
-		isCustomCity: 1,
-		lastSyncDate: new Date(),
-		latitude: 40.85216,
-		longitude: 14.26811,
-		name: 'adopted-probe-2',
-		network: 'IRIDEOS S.P.A.',
-		onlineTimesToday: 120,
-		state: null,
-		status: 'ready',
-		tags: JSON.stringify([{ value: 'tag-1', prefix: user.github_username }]),
-		systemTags: JSON.stringify([ 'datacenter-network' ]),
-		userId: user.id,
-		uuid: '681023cb-6aec-45a1-adde-e705c4043549',
-		version: '0.28.0',
-		hardwareDevice: null,
-	},
-	{
-		id: randomUUID(),
-		asn: 61493,
-		city: 'Buenos Aires',
-		country: 'AR',
-		countryOfCustomCity: null,
-		date_created: '2024-02-22 11:02:12',
-		date_updated: null,
-		ip: '131.255.7.26',
-		altIps: JSON.stringify([ '2001:1000:130f:2000:3000:09c0:876a:130b' ]),
-		isCustomCity: 0,
-		lastSyncDate: new Date(),
-		latitude: -34.6131,
-		longitude: -58.3772,
-		name: null,
-		network: 'InterBS S.R.L. (BAEHOST)',
-		onlineTimesToday: 0,
-		state: null,
-		status: 'ready',
-		tags: '[]',
-		systemTags: JSON.stringify([ 'datacenter-network' ]),
-		userId: user.id,
-		uuid: 'b42c4319-6be3-46d4-8a01-d4558f0c070c',
-		version: '0.28.0',
-		hardwareDevice: null,
-	},
-	{
-		id: randomUUID(),
-		asn: 16019,
-		city: 'Prague',
-		country: 'CZ',
-		countryOfCustomCity: null,
-		date_created: '2024-02-22 11:02:12',
-		date_updated: null,
-		ip: '2001:1000:130f:2000:3000:09c0:876a:130c',
-		altIps: JSON.stringify([ '131.255.7.26', '2001:1000:130f:2000:3000:09c0:876a:130d' ]),
-		isCustomCity: 0,
-		lastSyncDate: new Date(),
-		latitude: 50.0736,
-		longitude: 14.4185,
-		name: null,
-		network: 'Vodafone Czech Republic a.s.',
-		systemTags: JSON.stringify([ 'eyeball-network' ]),
-		onlineTimesToday: 0,
-		state: null,
-		status: 'offline',
-		tags: '[]',
-		userId: user.id,
-		uuid: 'b42c4319-6be3-46d4-8a01-d4558f0c070d',
-		version: '0.28.0',
-		hardwareDevice: null,
-	}]);
-
-	await knex('sponsors').insert([{
-		date_created: '2024-02-22 11:48:00',
-		date_updated: null,
-		user_created: null,
-		user_updated: null,
-		github_id: user.external_identifier,
-		github_login: user.github_username,
-		last_earning_date: '2024-02-22 11:48:00',
-		monthly_amount: 5,
-	},
-	{
-		date_created: '2024-02-22 11:48:00',
-		date_updated: null,
-		user_created: null,
-		user_updated: null,
-		github_id: '6192491',
-		github_login: 'MartinKolarik',
-		last_earning_date: '2024-02-22 11:48:00',
-		monthly_amount: 100,
-	}]);
-
-	await knex('gp_credits_additions').insert([{
-		amount: 10000,
-		comment: 'One-time $50 sponsorship.',
-		consumed: 1,
-		date_created: '2024-01-05 11:46:22',
-		github_id: user.external_identifier,
-		user_updated: null,
-		adopted_probe: null,
-	},
-	{
-		amount: 1000,
-		comment: 'Recurring $5 sponsorship.',
-		consumed: 1,
-		date_created: '2024-02-05 11:46:22',
-		github_id: user.external_identifier,
-		user_updated: null,
-		adopted_probe: null,
-	},
-	{
-		amount: 150,
-		comment: 'Adopted probe "adopted-probe-2" (213.136.174.80).',
-		consumed: 1,
-		date_created: '2024-03-05 11:46:22',
-		github_id: user.external_identifier,
-		user_updated: null,
-		adopted_probe: probeId,
-	},
-	{
-		amount: 150,
-		comment: 'Adopted probe "adopted-probe-2" (213.136.174.80).',
-		consumed: 1,
-		date_created: '2024-03-06 11:46:22',
-		github_id: user.external_identifier,
-		user_updated: null,
-		adopted_probe: probeId,
-	},
-	{
-		amount: 150,
-		comment: 'Adopted probe "adopted-probe-2" (213.136.174.80).',
-		consumed: 1,
-		date_created: '2024-03-07 11:46:22',
-		github_id: user.external_identifier,
-		user_updated: null,
-		adopted_probe: probeId,
-	},
-	{
-		amount: 150,
-		comment: 'Adopted probe "adopted-probe-2" (213.136.174.80).',
-		consumed: 1,
-		date_created: '2024-03-08 11:46:22',
-		github_id: user.external_identifier,
-		user_updated: null,
-		adopted_probe: probeId,
-	}, {
-		amount: 2000,
-		comment: 'One-time $10 sponsorship.',
-		consumed: 1,
-		date_created: '2024-04-05 11:46:22',
-		github_id: user.external_identifier,
-		user_updated: null,
-		adopted_probe: null,
-	},
-	{
-		amount: 3000,
-		comment: 'Recurring $15 sponsorship.',
-		consumed: 1,
-		date_created: '2024-05-05 11:46:22',
-		github_id: user.external_identifier,
-		user_updated: null,
-		adopted_probe: null,
-	}]);
-
-	await knex('gp_credits').where({ user_id: user.id }).update({ amount: knex.raw('amount - ?', [ 6000 ]) });
-
-	await knex('gp_credits_deductions').where({ user_id: user.id }).update({ date: '2024-04-01' });
-
-	await knex('gp_credits').where({ user_id: user.id }).update({ amount: knex.raw('amount - ?', [ 5000 ]) });
 };
